Separate sync to modules
