+++ crates/papyrus_sync/src/lib.rs
@@ -1,28 +1,22 @@
-#[cfg(test)]
-#[path = "sync_test.rs"]
-mod sync_test;
-
+mod block;
 mod sources;
+mod state;
 
 use std::sync::Arc;
 use std::time::Duration;
 
-use async_stream::stream;
-use futures_util::{pin_mut, select, Stream, StreamExt};
+use futures_util::{pin_mut, select, StreamExt};
 use indexmap::IndexMap;
-use papyrus_storage::body::{BodyStorageReader, BodyStorageWriter};
-use papyrus_storage::header::{HeaderStorageReader, HeaderStorageWriter};
-use papyrus_storage::ommer::{OmmerStorageReader, OmmerStorageWriter};
-use papyrus_storage::state::{StateStorageReader, StateStorageWriter};
-use papyrus_storage::{StorageError, StorageReader, StorageWriter, TransactionIndex};
+use papyrus_storage::{StorageError, StorageReader, StorageWriter};
 use serde::{Deserialize, Serialize};
 use starknet_api::block::{Block, BlockHash, BlockNumber};
 use starknet_api::core::ClassHash;
 use starknet_api::state::{ContractClass, StateDiff};
-use starknet_api::transaction::TransactionOffsetInBlock;
-use tracing::{debug, error, info, trace, warn};
+use tracing::{error, info, warn};
 
 pub use self::sources::{CentralError, CentralSource, CentralSourceConfig, CentralSourceTrait};
+use crate::block::{handle_block_reverts, store_block, stream_new_blocks};
+use crate::state::{store_state_diff, stream_new_state_diffs};
 
 #[derive(Clone, Copy, Serialize, Deserialize)]
 pub struct SyncConfig {
@@ -137,7 +131,8 @@ impl<TCentralSource: CentralSourceTrait + Sync + Send + 'static> GenericStateSyn
     //  2. Create infinite block and state diff streams to fetch data from the central source.
     //  3. Fetch data from the streams with unblocking wait while there is no new data.
     async fn sync_while_ok(&mut self) -> StateSyncResult {
-        self.handle_block_reverts().await?;
+        let txn = self.writer.begin_rw_txn()?;
+        handle_block_reverts(self.reader.clone(), txn, self.central_source.clone()).await?;
         let block_stream = stream_new_blocks(
             self.reader.clone(),
             self.central_source.clone(),
@@ -166,292 +161,25 @@ impl<TCentralSource: CentralSourceTrait + Sync + Send + 'static> GenericStateSyn
 
     // Tries to store the incoming data.
     async fn process_sync_event(&mut self, sync_event: SyncEvent) -> StateSyncResult {
+        let txn = self.writer.begin_rw_txn()?;
         match sync_event {
             SyncEvent::BlockAvailable { block_number, block } => {
-                self.store_block(block_number, block)
+                store_block(self.reader.clone(), txn, block_number, block)
             }
             SyncEvent::StateDiffAvailable {
                 block_number,
                 block_hash,
                 state_diff,
                 deployed_contract_class_definitions,
-            } => self.store_state_diff(
+            } => store_state_diff(
+                self.reader.clone(),
+                txn,
                 block_number,
                 block_hash,
                 state_diff,
                 deployed_contract_class_definitions,
             ),
         }
-    }
-
-    fn store_block(&mut self, block_number: BlockNumber, block: Block) -> StateSyncResult {
-        // Assuming the central source is trusted, detect reverts by comparing the incoming block's
-        // parent hash to the current hash.
-        self.verify_parent_block_hash(block_number, &block)?;
-
-        debug!("Storing block {block_number} with hash {}.", block.header.block_hash);
-        trace!("Block data: {block:#?}");
-        self.writer
-            .begin_rw_txn()?
-            .append_header(block_number, &block.header)?
-            .append_body(block_number, block.body)?
-            .commit()?;
-        Ok(())
-    }
-
-    fn store_state_diff(
-        &mut self,
-        block_number: BlockNumber,
-        block_hash: BlockHash,
-        state_diff: StateDiff,
-        deployed_contract_class_definitions: IndexMap<ClassHash, ContractClass>,
-    ) -> StateSyncResult {
-        if !self.is_reverted_state_diff(block_number, block_hash)? {
-            debug!("Storing state diff of block {block_number} with hash {block_hash}.");
-            trace!("StateDiff data: {state_diff:#?}");
-            self.writer
-                .begin_rw_txn()?
-                .append_state_diff(block_number, state_diff, deployed_contract_class_definitions)?
-                .commit()?;
-
-            // Info the user on syncing the block once all the data is stored.
-            info!("Added block {} with hash {}.", block_number, block_hash);
-        } else {
-            debug!(
-                "Storing ommer state diff of block {} with hash {:?}.",
-                block_number, block_hash
-            );
-            self.writer
-                .begin_rw_txn()?
-                .insert_ommer_state_diff(
-                    block_hash,
-                    &state_diff.into(),
-                    &deployed_contract_class_definitions,
-                )?
-                .commit()?;
-        }
-        Ok(())
-    }
-
-    // Compares the block's parent hash to the stored block.
-    fn verify_parent_block_hash(
-        &self,
-        block_number: BlockNumber,
-        block: &Block,
-    ) -> StateSyncResult {
-        let prev_block_number = match block_number.prev() {
-            None => return Ok(()),
-            Some(bn) => bn,
-        };
-        let prev_hash = self
-            .reader
-            .begin_ro_txn()?
-            .get_block_header(prev_block_number)?
-            .ok_or(StorageError::DBInconsistency {
-                msg: format!(
-                    "Missing block {prev_block_number} in the storage (for verifying block \
-                     {block_number}).",
-                ),
-            })?
-            .block_hash;
-
-        if prev_hash != block.header.parent_hash {
-            return Err(StateSyncError::ParentBlockHashMismatch {
-                block_number,
-                expected_parent_block_hash: block.header.parent_hash,
-                stored_parent_block_hash: prev_hash,
-            });
-        }
-
-        Ok(())
-    }
-
-    // Reverts data if needed.
-    async fn handle_block_reverts(&mut self) -> Result<(), StateSyncError> {
-        let header_marker = self.reader.begin_ro_txn()?.get_header_marker()?;
-
-        // Revert last blocks if needed.
-        let mut last_block_in_storage = header_marker.prev();
-        while let Some(block_number) = last_block_in_storage {
-            if self.should_revert_block(block_number).await? {
-                info!("Reverting block {}.", block_number);
-                self.revert_block(block_number)?;
-                last_block_in_storage = block_number.prev();
-            } else {
-                break;
-            }
-        }
-        Ok(())
-    }
-
-    // Deletes the block data from the storage, moving it to the ommer tables.
-    #[allow(clippy::expect_fun_call)]
-    fn revert_block(&mut self, block_number: BlockNumber) -> StateSyncResult {
-        // TODO: Modify revert functions so they return the deleted data, and use it for inserting
-        // to the ommer tables.
-        let mut txn = self.writer.begin_rw_txn()?;
-        let header = txn
-            .get_block_header(block_number)?
-            .expect(format!("Tried to revert a missing header of block {block_number}").as_str());
-        let transactions = txn.get_block_transactions(block_number)?.expect(
-            format!("Tried to revert a missing transactions of block {block_number}").as_str(),
-        );
-        let transaction_outputs = txn.get_block_transaction_outputs(block_number)?.expect(
-            format!("Tried to revert a missing transaction outputs of block {block_number}")
-                .as_str(),
-        );
-
-        // TODO: use iter_events of EventsReader once it supports RW transactions.
-        let mut events: Vec<_> = vec![];
-        for idx in 0..transactions.len() {
-            let tx_idx = TransactionIndex(block_number, TransactionOffsetInBlock(idx));
-            events.push(txn.get_transaction_events(tx_idx)?.unwrap_or_default());
-        }
-
-        txn = txn
-            .revert_header(block_number)?
-            .insert_ommer_header(header.block_hash, &header)?
-            .revert_body(block_number)?
-            .insert_ommer_body(
-                header.block_hash,
-                &transactions,
-                &transaction_outputs,
-                events.as_slice(),
-            )?;
-
-        let (txn, maybe_deleted_data) = txn.revert_state_diff(block_number)?;
-        if let Some((thin_state_diff, declared_classes)) = maybe_deleted_data {
-            txn.insert_ommer_state_diff(header.block_hash, &thin_state_diff, &declared_classes)?
-                .commit()?;
-        } else {
-            txn.commit()?;
-        }
-        Ok(())
-    }
-
-    /// Checks if centrals block hash at the block number is different from ours (or doesn't exist).
-    /// If so, a revert is required.
-    async fn should_revert_block(&self, block_number: BlockNumber) -> Result<bool, StateSyncError> {
-        if let Some(central_block_hash) = self.central_source.get_block_hash(block_number).await? {
-            let storage_block_header =
-                self.reader.begin_ro_txn()?.get_block_header(block_number)?;
-
-            match storage_block_header {
-                Some(block_header) => Ok(block_header.block_hash != central_block_hash),
-                None => Ok(false),
-            }
-        } else {
-            // Block number doesn't exist in central, revert.
-            Ok(true)
-        }
-    }
-
-    fn is_reverted_state_diff(
-        &self,
-        block_number: BlockNumber,
-        block_hash: BlockHash,
-    ) -> Result<bool, StateSyncError> {
-        let txn = self.reader.begin_ro_txn()?;
-        let storage_header = txn.get_block_header(block_number)?;
-        match storage_header {
-            Some(storage_header) if storage_header.block_hash == block_hash => Ok(false),
-            _ => {
-                // No matching header, check in the ommer headers.
-                match txn.get_ommer_header(block_hash)? {
-                    Some(_) => Ok(true),
-                    None => Err(StateSyncError::StateDiffWithoutMatchingHeader {
-                        block_number,
-                        block_hash,
-                    }),
-                }
-            }
-        }
-    }
-}
-
-fn stream_new_blocks<TCentralSource: CentralSourceTrait + Sync + Send>(
-    reader: StorageReader,
-    central_source: Arc<TCentralSource>,
-    block_propation_sleep_duration: Duration,
-) -> impl Stream<Item = Result<SyncEvent, StateSyncError>> {
-    stream! {
-        loop {
-            let header_marker = reader.begin_ro_txn()?.get_header_marker()?;
-            let last_block_number = central_source.get_block_marker().await?;
-            debug!("Downloading blocks [{} - {}).", header_marker, last_block_number);
-            if header_marker == last_block_number {
-                tokio::time::sleep(block_propation_sleep_duration).await;
-                continue;
-            }
-            let block_stream =
-                central_source.stream_new_blocks(header_marker, last_block_number).fuse();
-            pin_mut!(block_stream);
-            while let Some(maybe_block) = block_stream.next().await {
-                match maybe_block {
-                    Ok((block_number, block)) => yield Ok(SyncEvent::BlockAvailable { block_number, block }),
-                    Err(err) => {
-                        yield Err(StateSyncError::CentralSourceError(err));
-                        break;
-                    }
-                }
-            }
-        }
-    }
-}
-
-fn stream_new_state_diffs<TCentralSource: CentralSourceTrait + Sync + Send>(
-    reader: StorageReader,
-    central_source: Arc<TCentralSource>,
-    block_propation_sleep_duration: Duration,
-) -> impl Stream<Item = Result<SyncEvent, StateSyncError>> {
-    stream! {
-        loop {
-            let txn = reader.begin_ro_txn()?;
-            let state_marker = txn.get_state_marker()?;
-            let last_block_number = txn.get_header_marker()?;
-            drop(txn);
-            debug!("Downloading state diffs [{} - {}).", state_marker, last_block_number);
-            if state_marker == last_block_number {
-                tokio::time::sleep(block_propation_sleep_duration).await;
-                continue;
-            }
-            let state_diff_stream =
-                central_source.stream_state_updates(state_marker, last_block_number).fuse();
-            pin_mut!(state_diff_stream);
-
-            while let Some(maybe_state_diff) = state_diff_stream.next().await {
-                match maybe_state_diff {
-                    Ok((
-                        block_number,
-                        block_hash,
-                        mut state_diff,
-                        deployed_contract_class_definitions,
-                    )) => {
-                        sort_state_diff(&mut state_diff);
-                        yield Ok(SyncEvent::StateDiffAvailable {
-                            block_number,
-                            block_hash,
-                            state_diff,
-                            deployed_contract_class_definitions,
-                        });
-                    }
-                    Err(err) => {
-                        yield Err(StateSyncError::CentralSourceError(err));
-                        break;
-                    }
-                }
-            }
-        }
-    }
-}
-
-pub fn sort_state_diff(diff: &mut StateDiff) {
-    diff.declared_classes.sort_unstable_keys();
-    diff.deployed_contracts.sort_unstable_keys();
-    diff.nonces.sort_unstable_keys();
-    diff.storage_diffs.sort_unstable_keys();
-    for storage_entries in diff.storage_diffs.values_mut() {
-        storage_entries.sort_unstable_keys();
     }
 }
